+++ src/rust/src/utils/noise.rs
@@ -8,8 +8,8 @@ pub fn simplex(frequency: f32) -> Vec<f32> {
     .generate_scaled(-1.0, 1.0)
 }
 
-pub fn gaussian_buffer(size: usize) -> Vec<f32> {
-  let mut samples: Vec<f32> = Vec::with_capacity(size);
+pub fn gaussian_buffer(size: usize) -> Vec<f64> {
+  let mut samples: Vec<f64> = Vec::with_capacity(size);
   samples.fill_with(gaussian);
   // for _n in 0..size {
   //   samples.push(gaussian());
